+++ lib/artifact/plugins/push.rb
@@ -32,6 +32,7 @@ module Artifact
       plugin_argument :gpg_id, description: 'gpg id for encrypting'
       plugin_argument :gpg_passphrase, description: 'gpg passphrase', validator: Proc.new {|x| not x.nil? and not x.empty?}
 
+      plugin_argument :exclude, description: 'exclude files and directories', optional: true, default: nil
       plugin_argument :sign, description: 'sign pushed artifact', optional: true, default: false
       plugin_argument :signer, description: 'gpg id for signing', optional: true, default: nil
 
@@ -48,7 +49,11 @@ module Artifact
       # find files for artifact
       def find_artifact
         Dir.chdir "#{@workspace}/#{@target_directory}"
-        @artifacts = Find.find('./').find_all{|x| x =~ /#{@file_name_pattern}/}
+        @artifacts = Find.find('./').find_all do |filename|
+          filename =~ /#{@file_name_pattern}/ and
+          not (@exclude and @exclude.find{|exclude| filename =~ /#{exclude}/})
+        end
+        v 'Found following artifacts: ' + @artifacts.inspect
       end
 
       # archive files for artifact
@@ -57,33 +62,49 @@ module Artifact
         @data = Zip::OutputStream.write_buffer do |out|
           if @artifacts.length == 1
             file = @artifacts.shift
+            v '+ ' + File.basename(file)
             out.put_next_entry(File.basename(file))
             out.write File.read(file)
           else
             @artifacts.each do |artifact|
               next if artifact == './'
 
+              v '+ ' + artifact
               out.put_next_entry(artifact)
-              if not File.directory? artifact
+              if File.directory? artifact
+                artifact += '/' unless artifact.end_with? '/'
+                Zip::Entry.new(nil, artifact).write_to_zip_output_stream(out)
+              else
+                d 'w ' + artifact
                 out.write File.read(artifact)
               end
             end
           end
         end
         @data.rewind
+        d "Zip::OutputStream length: #{@data.length}"
+
+        raise 'ZIP data is empty' if @data.length == 0
       end
 
       # encrypt artifact
       def encrypt_artifact
         subsection 'Encrypt artifact', color: :green, prefix: @output_prefix unless @silent
-        crypto = GPGME::Crypto.new pinentry_mode: GPGME::PINENTRY_MODE_LOOPBACK, password: @gpg_passphrase, recipients: @gpg_id
+
+        crypto = GPGME::Crypto.new recipients: @gpg_id
         @data = crypto.encrypt(@data, always_trust: true)
+
+        d "GPGME::Crypto length: #{@data.to_s.length}"
+        raise 'GPG data is empty' if @data.to_s.length == 0
 
         if @sign
           subsection 'Sign artifact', color: :green, prefix: @output_prefix unless @silent
-          @sign_data = crypto.sign(@data, pinentry_mode: GPGME::PINENTRY_MODE_LOOPBACK, mode: GPGME::SIG_MODE_DETACH, password: @gpg_passphrase, signer: @signer)
+          @sign_data = crypto.sign(@data, mode: GPGME::SIG_MODE_DETACH, signer: @signer, pinentry_mode: GPGME::PINENTRY_MODE_LOOPBACK, password: @gpg_passphrase)
           @data.seek 0
         end
+        d "GPGME::Crypto length: #{@data.to_s.length}"
+
+        raise 'GPG data is empty' if @data.to_s.length == 0
       rescue
         raise
       end