+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-0.1.2 (UNRELEASED)
+0.1.3 (2018-09-13)
+==================
+
+- added fallback to popen for gpg 2.0 for --get
+- added verbose output
+- added exclude functionality to --push
+
+0.1.2 (2018-09-11)
 ==================
 
 - support for directories